v3.1

+++ cliente/cliente.py
@@ -7,9 +7,10 @@ import random
 import hashlib
 import os
 import time
+import os.path
 
 def joinFiles(msg,nombre):
-	with open(nombre,'wb') as file:
+	with open(nombre+'.png','wb') as file:
 		for x in msg['name']:
 			name = msg['name'][x]['namePart']
 			with open(name, 'rb') as f:
@@ -22,8 +23,8 @@ def upload(msg,identity):
 	for keys in name:
 		name= keys
 	for x in msg[name]:
-		print(msg[name][x]['servidor'])
-		print(msg[name][x]['namePart'])		
+		#print(msg[name][x]['servidor'])
+		#print(msg[name][x]['namePart'])		
 		context = zmq.Context()
 		socket = context.socket(zmq.DEALER)
 		socket.identity = identity
@@ -41,6 +42,7 @@ def upload(msg,identity):
 
 def download(msg,identity):
 	
+	print(msg)
 	for x in msg['name']:
 		context = zmq.Context()
 		socket = context.socket(zmq.DEALER)
@@ -56,33 +58,37 @@ def download(msg,identity):
 		time.sleep(1)
 		mensaje_json = json.loads(mensaje)
 		#operacion = mensaje_json['operacion']
-		print(mensaje_json['archivo']['nombre'])
-		print('-------')
+		#print(mensaje_json['archivo']['nombre'])
+		#print('-------')
 		with open(mensaje_json['archivo']['nombre'], 'wb') as f:
 			f.write(data)
 	
-	joinFiles(msg,'holamundo.png')	
+	joinFiles(msg,msg['nombreHash'])	
 
 def hashearArchivo(FILE):
-	SizeFile = os.stat(FILE).st_size
-	SizePart = 1024*10
-	diccionario = {}
-	DicPart = {}
-	FileComplete = hashlib.sha1()
-	with open(FILE, 'rb') as f:
-		i= 1
-		for x in range(int(SizeFile/SizePart)+1):			
-			data = f.read(SizePart)
-			objetohash = hashlib.sha1(data)
-			cadena = objetohash.hexdigest()
-			DicPart.update({i:{'namePart':cadena}})
-			FileComplete.update(data)
-			with open(cadena, 'wb') as part:
-				part.write(data)
-			i=i+1
-	shaprincipal =	FileComplete.hexdigest()	
-	diccionario.update({shaprincipal:DicPart})
-	return diccionario
+	if(os.path.isfile(FILE)): 
+		SizeFile = os.stat(FILE).st_size
+		SizePart = 1024*10*1024
+		diccionario = {}
+		DicPart = {}
+		FileComplete = hashlib.sha1()
+		with open(FILE, 'rb') as f:
+			i= 1
+			for x in range(int(SizeFile/SizePart)+1):			
+				data = f.read(SizePart)
+				objetohash = hashlib.sha1(data)
+				cadena = objetohash.hexdigest()
+				DicPart.update({i:{'namePart':cadena}})
+				FileComplete.update(data)
+				with open(cadena, 'wb') as part:
+					part.write(data)
+				i=i+1
+		shaprincipal =	FileComplete.hexdigest()	
+		diccionario.update({shaprincipal:DicPart})
+		return diccionario
+	else:
+		print('archivo no existe en carpeta')
+		return -1
 
 def hashf(FILE):
 	with open(FILE, 'rb') as f:
@@ -113,6 +119,7 @@ def main():
 	poller.register(sys.stdin, zmq.POLLIN)
 	poller.register(socket, zmq.POLLIN)
 	while True:
+		print("opciones[u:upload,d:download N archivo]")
 		socks = dict(poller.poll())
 		mensaje = {'operacion':'sin operacion'}
 		mensaje_json = json.dumps(mensaje)
@@ -125,16 +132,25 @@ def main():
 				upload(mensaje_json['lista'],identity)
 			elif(operacion=='download'):
 				download(mensaje_json,identity)
+			elif(operacion=='listar'):
+				print(mensaje_json['list'].keys())
 
 		elif sys.stdin.fileno() in socks:
-			print("opciones[u:upload,d:download Narchivo]")
+			print("?")
 			command = input()
 			op, msg = command.split(' ', 1)
 			if(op=='u'):
-				mensaje = {'operacion':'upload','lista':hashearArchivo('pruebaupload.png')}
-				mensaje_json = json.dumps(mensaje)
+				if(hashearArchivo(msg)!= -1):
+					mensaje = {'operacion':'upload','lista':hashearArchivo(msg)}
+					mensaje_json = json.dumps(mensaje)
+				else:
+					print('no existe archivo')
+					pass
 			elif(op=='d'):
 				mensaje = {'operacion':'download','name':'aae5ae766cd09d09ff4fc98cc118d4a8c7bda4ff'}
+				mensaje_json = json.dumps(mensaje)
+			elif(op=='l'):
+				mensaje = {'operacion':'listar'}
 				mensaje_json = json.dumps(mensaje)
 			else:
 				pass
